backend-jombo-essaygrade/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pymysql==1.1.2
python-dotenv==1.2.2
pydantic==2.12.5
passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
cryptography==46.0.5